requirements.txt
fastapi>=0.111.1
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
asyncpg==0.29.0
fastapi_mail==1.4.1
sqlalchemy==1.4.25
